Added repo aggregation. Tests pass.
Added `global-repo` and `priority-repo`, which will take
arbitrary repo functions and aggregate their results.
`priority-repo` takes the first non-empty result from the
first repo in a given list and returns it. `global-repo` takes
the results of all repositories given and sorts them according to
a given comparator.

Also, I moved `map-query` to the file `src/degasolv/resolver_utils.clj`
as this function is not directly used by the function `resolve-dependencies`.